requirements.txt
flask
werkzeug
numpy
tensorflow
pillow
gunicorn